add a new hocs folder with toggler in it
